+++ src/graphics/draw/draw_circle.c
@@ -2,7 +2,7 @@
 
 static void		update_p_and_decision(int in_loop, t_ivec2 *p, int *decision,
 					int radius);
-static t_color	interpolate_color(t_color start, t_color end, double ratio);
+static t_color	interpolate_color(t_color start, t_color end, float ratio);
 
 /* Bresenham/Midpoint Circle Drawing algorithm */
 void	draw_circle(t_man *man, t_vert center, int radius)
@@ -66,8 +66,8 @@ void	draw_circle_full_gradient(t_man *man, t_vert center, int radius,
 {
 	t_ivec2	p;
 	int		distance_squared;
-	double	linear_ratio;
-	double	adjusted_ratio;
+	float	linear_ratio;
+	float	adjusted_ratio;
 	t_color	c;
 
 	p.y = -radius;
@@ -79,7 +79,7 @@ void	draw_circle_full_gradient(t_man *man, t_vert center, int radius,
 			distance_squared = p.x * p.x + p.y * p.y;
 			if (distance_squared <= radius * radius)
 			{
-				linear_ratio = sqrt_approx((double)distance_squared) / radius;
+				linear_ratio = sqrt_approx((float)distance_squared) / radius;
 				adjusted_ratio = linear_ratio * linear_ratio;
 				c = interpolate_color(center.color, edge, adjusted_ratio);
 				draw_point(man, c, center.coord.x + p.x, center.coord.y + p.y);
@@ -106,7 +106,7 @@ static void	update_p_and_decision(int in_loop, t_ivec2 *p, int *decision,
 	return ;
 }
 
-static t_color	interpolate_color(t_color start, t_color end, double ratio)
+static t_color	interpolate_color(t_color start, t_color end, float ratio)
 {
 	t_color	result;
