+++ src/resolution.c
@@ -2,7 +2,7 @@
 
 static t_ivec2	adapt_size_to_ratio(t_man *man, t_ivec2 size);
 
-int	set_resolution(t_man *man, t_ivec2 window_size, double aspect_ratio)
+int	set_resolution(t_man *man, t_ivec2 window_size, float aspect_ratio)
 {
 	man->res.monitor_size = get_monitor_size(man);
 	man->res.window_size_default.x = clamp(window_size.x,
@@ -12,10 +12,10 @@ int	set_resolution(t_man *man, t_ivec2 window_size, double aspect_ratio)
 	man->res.window_size.x = man->res.window_size_default.x;
 	man->res.window_size.y = man->res.window_size_default.y;
 	set_ivec2(&man->res.window_position, 0, 0);
-	man->res.half_fov_tan = tan(deg2rad(53) / 2.0);
+	man->res.half_fov_tan = tanf(deg2rad(53) / 2.0);
 	man->res.ratio = aspect_ratio;
 	if (aspect_ratio <= 0.1)
-		man->res.ratio = (double)man->res.monitor_size.x
+		man->res.ratio = (float)man->res.monitor_size.x
 			/ man->res.monitor_size.y;
 	if (aspect_ratio <= 0.1)
 		return (0);
